pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svn-ai-reviewer"
version = "1.0.0"
description = "SVN AI 智能审查助手 - 基于 AI 的 SVN 代码审查 CLI 工具"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "EDY"},
]
keywords = ["svn", "code-review", "ai", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0",
    "rich>=12.0",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.scripts]
svn-ai = "cli:main"

[tool.setuptools]
py-modules = [
    "cli",
    "svn_client",
    "config_manager",
    "prompt_builder",
    "report_generator",
    "batch_processor",
    "log_generator",
]
packages = [
    "commands",
    "models",
    "ai_provider",
    "templates",
]

[tool.setuptools.package-data]
templates = ["*.md"]
